+++ demo/shared/src/commonMain/kotlin/com/huanshankeji/compose/multiplatformhtmlunified/demo/App.kt
@@ -1,4 +1,4 @@
-package com.huanshankeji.compose.material.demo
+package com.huanshankeji.compose.multiplatformhtmlunified.demo
 
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.unit.dp