+++ demo/shared/src/commonMain/kotlin/com/huanshankeji/compose/multiplatformhtmlunified/demo/Material3ViewModel.kt
@@ -1,4 +1,4 @@
-package com.huanshankeji.compose.material.demo
+package com.huanshankeji.compose.multiplatformhtmlunified.demo
 
 import androidx.lifecycle.ViewModel
 import kotlinx.coroutines.flow.MutableStateFlow